requirements.txt
streamlit
langchain
langgraph
python-dotenv
requests
pypdf
python-docx
tiktoken
faiss-cpu
rapidfuzz
transformers